molminer/detection/requirements.txt
absl-py==2.1.0
aiobotocore==2.13.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
awscli==1.32.106
black==24.4.2
boto3==1.34.106
botocore==1.34.106
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
docutils==0.16
effdet==0.4.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fancycompleter==0.9.1
fastapi==0.83.0
filelock==3.14.0
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.6.0
fvcore==0.1.5.post20221221
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.23.2
hydra-core==1.3.2
idna==3.7
iopath==0.1.9
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
layoutparser==0.3.4
linkify-it-py==2.0.3
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pdf-annotate==0.12.0
pdf2image==1.16.2
pdfminer.six==20231228
pdfplumber==0.11.0
pdfrw==0.4
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.8.2
protobuf==5.27.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pybind11==2.12.0
pycocotools==2.0.7
pycparser==2.22
pydantic==1.10.15
Pygments==2.18.0
pyparsing==3.1.2
pypdfium2==4.30.0
pyrepl==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rdkit==2023.9.6
redun==0.19.5
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.7.2
s3fs==2024.6.0
s3transfer==0.10.1
safetensors==0.4.3
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.19.1
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
textual==0.65.1
thop==0.1.1.post2209072238
timm==1.0.3
tokenizers==0.19.1
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.66.4
transformers==4.41.2
triton==2.3.0
typing_extensions==4.12.0
tzdata==2024.1
uc-micro-py==1.0.3
ultralytics==8.2.25
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yacs==0.1.8
yarl==1.9.4
